+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Timeline rendering error in Browse OTU
 - Fix wildcard by attribute in Filter collection object
 - Confidences modal height in radial annotator [#2304]
+- Fix empty search in Filter collection objects
 
 
 [#2304]: https://github.com/SpeciesFileGroup/taxonworks/issues/2304